+++ test/src/Test/TestClassWithFinalReturnType.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Eloquent\Phony\Test;
+
+class TestClassWithFinalReturnType
+{
+    public function finalReturnType(): TestFinalClass
+    {
+        return new TestFinalClass();
+    }
+}