+++ test/src/Test/functions.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Eloquent\Phony\Test;
+
+function testFunctionWithFinalReturnType(): TestFinalClass
+{
+    return new TestFinalClass();
+}